unify can drop + Form editor prevent opening, when cannot drop
