+++ frontend/src/js/small-tab-navigation/HoverNavigatable.tsx
@@ -28,7 +28,6 @@ const Root = styled("div")<{
     return "inherit";
   }};
   position: relative;
-  border-radius: ${({ theme }) => theme.borderRadius};
   display: inline-flex;
 `;
